Use rename rather than copy for identical snapshot
When the snapshot for a deleted file would be identical to the file
itself and compression is disabled, simply rename rather than copying.
This can make deleted file handling much faster.

I have been using this for some years in my environment with no issues,
but there may be caveats, such as if somehow the backup crosses
filesystem boundaries.